requirements.txt
Django==2.1.7
django-js-asset==1.2.2
django-mptt==0.9.1
pytz==2018.9
